Make json an optional feature
